# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Talanta-Cloud Solutions EMS API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: employees
    description: Employee management
  - name: departments
    description: Department management
  - name: branches
    description: Branch/location management
  - name: dashboard
    description: Dashboard statistics
  - name: ai
    description: AI-powered features
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees with optional search/filter
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or email
        - name: department
          in: query
          schema:
            type: string
          description: Filter by department ID
        - name: branch
          in: query
          schema:
            type: string
          description: Filter by branch ID
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
          description: Filter by status
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get a single employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a new department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "201":
          description: Created department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"

  /departments/{id}:
    put:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List all branches/locations
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a new branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchBody"
      responses:
        "201":
          description: Created branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"

  /branches/{id}:
    put:
      operationId: updateBranch
      tags: [branches]
      summary: Update a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchBody"
      responses:
        "200":
          description: Updated branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get overall dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/employees-by-department:
    get:
      operationId: getEmployeesByDepartment
      tags: [dashboard]
      summary: Employee count grouped by department
      responses:
        "200":
          description: Employee counts by department
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupCount"

  /dashboard/employees-by-branch:
    get:
      operationId: getEmployeesByBranch
      tags: [dashboard]
      summary: Employee count grouped by branch
      responses:
        "200":
          description: Employee counts by branch
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupCount"

  /dashboard/recent-employees:
    get:
      operationId: getRecentEmployees
      tags: [dashboard]
      summary: Recently added employees
      responses:
        "200":
          description: Recent employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"

  /ai/generate-summary:
    post:
      operationId: generateEmployeeSummary
      tags: [ai]
      summary: Auto-generate a professional summary for an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateSummaryBody"
      responses:
        "200":
          description: Generated summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateSummaryResponse"

  /ai/smart-search:
    post:
      operationId: smartSearch
      tags: [ai]
      summary: AI-powered employee search with natural language
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmartSearchBody"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Employee:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        jobTitle:
          type: string
        departmentId:
          type: integer
        departmentName:
          type: string
        branchId:
          type: integer
        branchName:
          type: string
        phone:
          type: string
        email:
          type: string
        dateOfEmployment:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive]
        photoUrl:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - jobTitle
        - departmentId
        - branchId
        - phone
        - email
        - dateOfEmployment
        - status
        - createdAt
        - updatedAt

    CreateEmployeeBody:
      type: object
      properties:
        fullName:
          type: string
          minLength: 1
        jobTitle:
          type: string
          minLength: 1
        departmentId:
          type: integer
        branchId:
          type: integer
        phone:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        dateOfEmployment:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive]
          default: active
        photoUrl:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
      required:
        - fullName
        - jobTitle
        - departmentId
        - branchId
        - phone
        - email
        - dateOfEmployment

    UpdateEmployeeBody:
      type: object
      properties:
        fullName:
          type: string
        jobTitle:
          type: string
        departmentId:
          type: integer
        branchId:
          type: integer
        phone:
          type: string
        email:
          type: string
          format: email
        dateOfEmployment:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive]
        photoUrl:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true

    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        employeeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - employeeCount
        - createdAt

    CreateDepartmentBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
      required:
        - name

    Branch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
          nullable: true
        employeeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - employeeCount
        - createdAt

    CreateBranchBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        location:
          type: string
          nullable: true
      required:
        - name

    DashboardStats:
      type: object
      properties:
        totalEmployees:
          type: integer
        activeEmployees:
          type: integer
        totalDepartments:
          type: integer
        totalBranches:
          type: integer
        newThisMonth:
          type: integer
      required:
        - totalEmployees
        - activeEmployees
        - totalDepartments
        - totalBranches
        - newThisMonth

    GroupCount:
      type: object
      properties:
        name:
          type: string
        count:
          type: integer
      required:
        - name
        - count

    GenerateSummaryBody:
      type: object
      properties:
        fullName:
          type: string
        jobTitle:
          type: string
        department:
          type: string
        branch:
          type: string
        dateOfEmployment:
          type: string
      required:
        - fullName
        - jobTitle
        - department

    GenerateSummaryResponse:
      type: object
      properties:
        summary:
          type: string
      required:
        - summary

    SmartSearchBody:
      type: object
      properties:
        query:
          type: string
          minLength: 1
      required:
        - query

    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string